Added character class and minor updates

+++ Engine/Engine.cpp
@@ -99,7 +99,7 @@ void Engine::Update()
 
 void Engine::BeginRender()
 {
-  glClearColor((float)0.56, (float)0.4, (float)0.32, 1);  // Clear back buffer to brown (R, G, B, alpha) <- Floats (Divide RGB values by 255)
+  glClearColor(0.56f, 0.4f, 0.32f, 1);  // Clear back buffer to brown (R, G, B, alpha) <- Floats (Divide RGB values by 255)
   glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);  // Clear the depth
 }
